requirements.txt
python-docx>=1.1.0
openpyxl>=3.1.0
click>=8.1.0
pyyaml>=6.0
pytest>=7.4.0
